moving ApplicationsCalendar to into Features folder structure moving and Renaming App.css for consistency
