+++ webui/app-tracker/src/components/Pages/Home/Home.js
@@ -1,7 +1,7 @@
 import React from "react";
 import Menu from "../../Menu/Menu";
 import Summary from "../../Features/Summary/Summary";
-import ApplicationsCalendar from "../../ApplicationsCalendar";
+import ApplicationsCalendar from "../../Features/ApplicationsCalendar/ApplicationsCalendar";
 
 export default function Home() {
   return (
